pyproject.toml
[tool.poetry]
name = "sidus-task"
version = "0.1.0"
description = "The project for user auth setup demonstration."
authors = ["dshubenok <dshubenok93@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.37"
sqlmodel = "^0.0.6"
alembic = "^1.8.0"
asyncpg = "^0.25.0"
passlib = "^1.7.4"
PyJWT = "^2.4.0"
pydantic = {extras = ["email"], version = "^1.9.1"}
bcrypt = "^3.2.2"
python-multipart = "^0.0.5"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
httpx = "^0.23.0"
aioredis = "^2.0.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"
celery = "^5.2.7"
Jinja2 = "^3.1.2"
redis = "^4.3.3"
psycopg2-binary = "^2.9.3"
greenlet = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
